Tip Calculator

+++ Python-Projects/Tip_Calculator.py
@@ -3,7 +3,7 @@
 
 print("Welcome to the tip calculator!")
 
-total_bill = int(input("What was the total bill? $"))
+total_bill = float(input("What was the total bill? $"))
 
 tip = int(input("How much tip would you like to give? \n 10,12, or 15? "))
